Feature: busylight_core.light.Light.nleds property
It occurred to me that the number of addressable LEDs a light has is
not discoverable, so I fixed it. Added a property `nleds` to the Light
class with a default value of zero indicating no addressable lights
(the device presents as a single light). Updated Light subclasses that
support hardware with addressable LEDs:

- Agile Innovative BlinkStick family
- EPOS Busylight
- Luxafor Flag
- Luxafor Busy Tag
- ThingM Blink1