Feature: Acquire Light instances using path specifier.
Addressing comments in JnyJny/busylight#3

Modified classmethods:
- busylight_core.light.Light.all_lights
- busylight_core.light.Light.first_light

Added an optional method argument `predicate` which is expected to be
a callable whose argument is a busylight_core.hardware.Hardware
instance and returns a bool. The predicate provides a means to filter
hardware devices before the Light is instantiated and potentially
acquired. If the predicate is None, no filter is performed.

There is some risk here since there is no garantee that the USB
hardware will provide the values the predicate is filtering on.

Predicates should be written defensively to avoid exceptions.

Added classmethod:
- busylight_core.light.Light.at_path(path, reset, exclusive)

The at_path classmethod provides a means to acquire a Light instance
based on a path specifier. The path specifier is a string which
is operating system dependent. The reset and exclusive arguments
are bools which are passed to the underlying Light.__init__ method
to control hardware reset and exclusive access respectively.